Refactor Gnumeric Style Reader into a separate dedicated class
